optimized i/o ops in io.py added 32-filter option to hyper.py

+++ src/tf_mmciad/utils/custom_loss.py
@@ -98,7 +98,7 @@ def jaccard2_coef(y_true, y_pred, smooth=SMOOTH):
     :return: Jaccard coefficient
     :rtype: float
     """
-    y_true_f = K.flatten(y_true)
+    y_true_f = K.flatten(tf.cast(y_true, dtype=tf.float32))
     y_pred_f = K.flatten(y_pred)
     intersection = K.sum(y_true_f * y_pred_f)
     union = K.sum(y_true_f * y_true_f) + K.sum(y_pred_f * y_pred_f) - intersection
